test update view

+++ src/main/java/com/heroku/java/model/StaffModel.java
@@ -1,21 +1,20 @@
 package com.heroku.java.model;
 
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "staff")
 public class StaffModel {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "staffid")
-    private Long staffId;
-    
+    private Integer staffId;
+
     @Column(name = "staffname")
     private String staffName;
 
@@ -30,7 +29,7 @@ public class StaffModel {
     }
 
     // Parameterized constructor
-    public StaffModel(Long staffId, String staffName, String staffEmail, String staffPassword) {
+    public StaffModel(int staffId, String staffName, String staffEmail, String staffPassword) {
         this.staffId = staffId;
         this.staffName = staffName;
         this.staffEmail = staffEmail;
@@ -38,11 +37,11 @@ public class StaffModel {
     }
 
     // Getters and setters
-    public Long getStaffId() {
+    public Integer getStaffId() {
         return staffId;
     }
 
-    public void setStaffId(Long staffId) {
+    public void setStaffId(Integer staffId) {
         this.staffId = staffId;
     }
 
@@ -72,11 +71,11 @@ public class StaffModel {
 
     @Override
     public String toString() {
-        return "StaffModel{" +
-                "staffId=" + staffId +
-                ", staffName='" + staffName + '\'' +
-                ", staffEmail='" + staffEmail + '\'' +
-                ", staffPassword='" + staffPassword + '\'' +
-                '}';
+        return "StaffModel{"
+                + "staffId=" + staffId
+                + ", staffName='" + staffName + '\''
+                + ", staffEmail='" + staffEmail + '\''
+                + ", staffPassword='" + staffPassword + '\''
+                + '}';
     }
 }
